refactor(dashboard): remove orphaned StatsView, i18n relative time, extract format utils
