+++ webview-ui/src/i18n/locales/fr/dashboard.json
@@ -76,5 +76,12 @@
     "mode": "Mode",
     "time": "Heure",
     "status": "Statut"
+  },
+  "time": {
+    "justNow": "a l'instant",
+    "minutesAgo": "il y a {{count}} min",
+    "hoursAgo": "il y a {{count}} h",
+    "yesterday": "hier",
+    "daysAgo": "il y a {{count}} jours"
   }
 }
