+++ webview-ui/src/i18n/locales/id/dashboard.json
@@ -76,5 +76,12 @@
     "mode": "Mode",
     "time": "Waktu",
     "status": "Status"
+  },
+  "time": {
+    "justNow": "baru saja",
+    "minutesAgo": "{{count}} mnt lalu",
+    "hoursAgo": "{{count}} jam lalu",
+    "yesterday": "kemarin",
+    "daysAgo": "{{count}} hari lalu"
   }
 }
