+++ webview-ui/src/i18n/locales/ja/dashboard.json
@@ -76,5 +76,12 @@
     "mode": "モード",
     "time": "時刻",
     "status": "ステータス"
+  },
+  "time": {
+    "justNow": "たった今",
+    "minutesAgo": "{{count}}分前",
+    "hoursAgo": "{{count}}時間前",
+    "yesterday": "昨日",
+    "daysAgo": "{{count}}日前"
   }
 }
