+++ webview-ui/src/i18n/locales/zh-CN/dashboard.json
@@ -76,5 +76,12 @@
     "mode": "模式",
     "time": "时间",
     "status": "状态"
+  },
+  "time": {
+    "justNow": "刚刚",
+    "minutesAgo": "{{count}}分钟前",
+    "hoursAgo": "{{count}}小时前",
+    "yesterday": "昨天",
+    "daysAgo": "{{count}}天前"
   }
 }
